% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/appx.r
\name{AS269}
\alias{AS269}
\title{Higher order Cornish Fisher approximation.}
\usage{
AS269(z,cumul,order.max=NULL,all.ords=FALSE)
}
\arguments{
\item{z}{the quantiles of the normal distribution. an atomic vector.}

\item{cumul}{the standardized cumulants of order 3, 4, ..., k. an atomic
vector.}

\item{order.max}{the maximum order approximation, must be greater than
\code{length(cumul)+2}.
We assume the cumulants have been adjusted to reflect that the random
variable has unit variance ('standardized cumulants')}

\item{all.ords}{a logical value. If \code{TRUE}, then results are returned
as a matrix, with a column for each order of the approximation. Otherwise
the results are a matrix with a single column of the highest order
approximation.}
}
\value{
A matrix, which is, depending on \code{all.ords}, either with one column per
order of the approximation, or a single column giving the maximum order
approximation. There is one row per value in \code{z}.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Cornish Fisher approximate quantiles on a standardized random variable.
}
\details{
Really detailed. \eqn{\zeta}{zeta}.

A list:
\itemize{
\item I use \eqn{n}{n} to stand for blah.
\item and so forth....
}
}
\examples{
foo <- AS269(seq(-2,2,0.01),c(0,2,0,4))
# test with the normal distribution:
s.cumul <- c(0,0,0,0,0,0,0,0,0)
pv <- seq(0.001,0.999,0.001)
zv <- qnorm(pv)
apq <- AS269(zv,s.cumul,all.ords=FALSE)
err <- truq - apq

# test with the exponential distribution
rate <- 0.7
n <- 18
# these are 'raw' cumulants'
cumul <- (rate ^ -(1:n)) * factorial(0:(n-1))
# standardize and chop
s.cumul <- cumul[3:length(cumul)] / (cumul[2]^((3:length(cumul))/2))
pv <- seq(0.001,0.999,0.001)
zv <- qnorm(pv)
apq <- cumul[1] + sqrt(cumul[2]) * AS269(zv,s.cumul,all.ords=TRUE)
truq <- qexp(pv, rate=rate)
err <- truq - apq
colSums(abs(err))

# an example from Wikipedia page on CF, \\url{https://en.wikipedia.org/wiki/Cornish\%E2\%80\%93Fisher_expansion}
s.cumul <- c(5,2)
apq <- 10 + sqrt(25) * AS269(qnorm(0.95),s.cumul,all.ords=TRUE)
w <- AS269(qnorm(0.95),s.cumul)
qv <- 10 + 5 * w[1,2]
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
Lee, Y-S., and Lin, T-K. "Algorithm AS269: High Order Cornish Fisher
Expansion." Appl. Stat. 41, no. 1 (1992): 233-240.
\url{http://www.jstor.org/stable/2347649}

Lee, Y-S., and Lin, T-K. "Correction to Algorithm AS269: High Order
Cornish Fisher Expansion." Appl. Stat. 42, no. 1 (1993): 268-269.
\url{http://www.jstor.org/stable/2347433}

Jaschke, Stefan R. "The Cornish-Fisher-expansion in the context of
Delta-Gamma-normal approximations." No. 2001, 54. Discussion Papers,
Interdisciplinary Research Project 373: Quantification and Simulation of
Economic Processes, 2001.
\url{http://www.jaschke-net.de/papers/CoFi.pdf}

AS 269. \url{http://lib.stat.cmu.edu/apstat/269}
}
\keyword{distribution}

