% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nakagami.r
\name{nakagami}
\alias{dnakagami}
\alias{nakagami}
\alias{pnakagami}
\alias{qnakagami}
\alias{rnakagami}
\title{The Nakagami distribution.}
\usage{
dnakagami(x, m, Omega=1, log = FALSE)

pnakagami(q, m, Omega=1, lower.tail = TRUE, log.p = FALSE)

qnakagami(p, m, Omega=1, lower.tail = TRUE, log.p = FALSE)

rnakagami(n, m, Omega=1)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{m}{the shape parameter.}

\item{Omega}{the spread parameter.}

\item{log}{logical; if TRUE, densities \eqn{f} are given
as \eqn{\mbox{log}(f)}{log(f)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log.p}{logical; if TRUE, probabilities p are given
as \eqn{\mbox{log}(p)}{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
 \eqn{P[X \le x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}
}
\value{
\code{dnakagami} gives the density, \code{pnakagami} gives the
distribution function, \code{qnakagami} gives the quantile function,
and \code{rnakagami} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Density, distribution function, quantile function and random
generation for the Nakagami distribution.
}
\details{
Let \eqn{Y \sim \mathrm{Gamma}\left(m, \Omega/m\right)}{Y ~ Gamma(m,Omega/m)}.
Then \eqn{\sqrt{Y}}{sqrt(Y)} takes a \emph{Nakagami distribution}
with shape \eqn{m}{m} and spread \eqn{\Omega}{Omega}.
Equivalently, a Nakagami is a rescaled chi random variable.
}
\note{
This is a thin wrapper on the gamma distribution functions.
}
\examples{
rvs <- rnakagami(128, 20, 1)
dvs <- dnakagami(rvs, 20, 1)
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
M. Nakagami. "The m-Distribution, a general formula of intensity of rapid fading."
Statistical Methods in Radio Wave Propagation (1960): pp 3-36.
}
\seealso{
gamma distribution functions, \code{\link{dgamma}, \link{pgamma}, \link{qgamma}, \link{rgamma}},
the chi distribution functions, \code{\link{dchi}, \link{pchi}, \link{qchi}, \link{rchi}}
}
\keyword{distribution}

