% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumchisq.r
\name{sumchisq}
\alias{dsumchisq}
\alias{psumchisq}
\alias{qsumchisq}
\alias{rsumchisq}
\alias{sumchisq}
\title{The sum of (non-central) chi-squares distribution.}
\usage{
dsumchisq(x, wts, df, ncp=0, order.max=6, log = FALSE)

psumchisq(q, wts, df, ncp=0, order.max=6, lower.tail = TRUE, log.p = FALSE)

qsumchisq(p, wts, df, ncp=0, order.max=6, lower.tail = TRUE, log.p = FALSE)

rsumchisq(n, wts, df, ncp=0)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{wts}{the vector of weights. We do \emph{not} vectorize over
\code{wts}, except against \code{df} and \code{ncp}.}

\item{df}{the vector of degrees of freedom. We do \emph{not} vectorize over
\code{df}, except against \code{wts} and \code{ncp}.}

\item{ncp}{the vector of non-centrality parameters. We do \emph{not} vectorize over
\code{ncp}, except against \code{df} and \code{wts}.}

\item{order.max}{the order to use in the approximate density,
distribution, and quantile computations, via the Gram-Charlier,
Edeworth, or Cornish-Fisher expansion.}

\item{log}{logical; if TRUE, densities \eqn{f} are given
as \eqn{\mbox{log}(f)}{log(f)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log.p}{logical; if TRUE, probabilities p are given
as \eqn{\mbox{log}(p)}{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
 \eqn{P[X \le x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}
}
\value{
\code{dsumchisq} gives the density, \code{psumchisq} gives the
distribution function, \code{qsumchisq} gives the quantile function,
and \code{rsumchisq} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Density, distribution function, quantile function and random
generation for the distribution of the weighted sum of non-central
chi-squares.
}
\details{
Let \eqn{X_i \sim \chi^2\left(\delta_i, \nu_i\right)}{X_i ~ chi^2(delta_i, v_i)}
be independently distributed non-central chi-squares, where \eqn{\nu_i}{v_i}
are the degrees of freedom, and \eqn{\delta_i}{delta_i} are the
non-centrality parameters.
Let \eqn{w_i} be given constants. Suppose
\deqn{Y = \sum_i w_i X_i.}{Y = sum w_i X_i.}
Then \eqn{Y}{Y} follows a weighted sum of chi-squares distribution. When
the weights are all one, and the chi-squares are all central, then
\eqn{Y}{Y} also follows a chi-square distribution.
}
\note{
The PDF, CDF, and quantile function are approximated, via
the Edgeworth or Cornish Fisher approximations, which may
not be terribly accurate in the tails of the distribution.
You are warned.

The distribution parameters are \emph{not} recycled
with respect to the \code{x, p, q} or \code{n} parameters,
for, respectively, the density, distribution, quantile
and generation functions. This is for simplicity of
implementation and performance. It is, however, in contrast
to the usual R idiom for dpqr functions.
}
\examples{
wts <- c(1,-3,4)
df <- c(100,20,10)
ncp <- c(5,3,1)
rvs <- rsumchisq(128, wts, df, ncp)
dvs <- dsumchisq(rvs, wts, df, ncp)
qvs <- psumchisq(rvs, wts, df, ncp)
pvs <- qsumchisq(ppoints(length(rvs)), wts, df, ncp)
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\seealso{
chi-square distribution functions,
\code{\link{dchisq}, \link{pchisq}, \link{qchisq}, \link{rchisq}},
}
\keyword{distribution}

