% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumchisqpow.r
\name{sumchisqpow}
\alias{sumchisqpow}
\alias{dsumchisqpow}
\alias{psumchisqpow}
\alias{qsumchisqpow}
\alias{rsumchisqpow}
\title{The sum of (non-central) chi-squares raised to powers distribution.}
\usage{
dsumchisqpow(x, wts, df, ncp=0, pow=1, log = FALSE, order.max=6)

psumchisqpow(q, wts, df, ncp=0, pow=1, lower.tail = TRUE, log.p = FALSE, order.max=6)

qsumchisqpow(p, wts, df, ncp=0, pow=1, lower.tail = TRUE, log.p = FALSE, order.max=6)

rsumchisqpow(n, wts, df, ncp=0, pow=1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{wts}{the vector of weights. 
This is recycled against the \code{df, ncp, pow}, but not against the \code{x,q,p,n}.}

\item{df}{the vector of degrees of freedom. 
This is recycled against the \code{wts, ncp, pow}, but not against the \code{x,q,p,n}.}

\item{ncp}{the vector of non-centrality parameters. 
This is recycled against the \code{wts, df, pow}, but not against the \code{x,q,p,n}.}

\item{pow}{the vector of the power parameters. 
This is recycled against the \code{wts, df, ncp}, but not against the \code{x,q,p,n}.}

\item{log}{logical; if TRUE, densities \eqn{f} are given 
as \eqn{\mbox{log}(f)}{log(f)}.}

\item{order.max}{the order to use in the approximate density,
distribution, and quantile computations, via the Gram-Charlier,
Edeworth, or Cornish-Fisher expansion.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log.p}{logical; if TRUE, probabilities p are given 
as \eqn{\mbox{log}(p)}{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}
}
\value{
\code{dsumchisqpow} gives the density, \code{psumchisqpow} gives the 
distribution function, \code{qsumchisqpow} gives the quantile function, 
and \code{rsumchisqpow} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Density, distribution function, quantile function and random
generation for the distribution of the weighted sum of non-central
chi-squares taken to powers.
}
\details{
Let \eqn{X_i \sim \chi^2\left(\delta_i, \nu_i\right)}{X_i ~ chi^2(delta_i, v_i)}
be independently distributed non-central chi-squares, where \eqn{\nu_i}{v_i}
are the degrees of freedom, and \eqn{\delta_i}{delta_i} are the
non-centrality parameters.  
Let \eqn{w_i} and \eqn{p_i} be given constants. Suppose
\deqn{Y = \sum_i w_i X_i^{p_i}.}{Y = sum w_i (X_i)^(p_i).}
Then \eqn{Y}{Y} follows a weighted sum of chi-squares to power distribution.
}
\note{
The PDF, CDF, and quantile function are approximated, via
the Edgeworth or Cornish Fisher approximations, which may
not be terribly accurate in the tails of the distribution.
You are warned.

The distribution parameters are \emph{not} recycled
with respect to the \code{x, p, q} or \code{n} parameters,
for, respectively, the density, distribution, quantile
and generation functions. This is for simplicity of
implementation and performance. It is, however, in contrast
to the usual R idiom for dpqr functions.

The 'sum of chisquare power' distribution does \emph{not} generalize
the 'chi-bar-square' distribution, whose \emph{density} is the sum of
chi-square densities.
}
\examples{
wts <- c(1,-3,4)
df <- c(100,20,10)
ncp <- c(5,3,1)
pow <- c(1,0.5,1)
rvs <- rsumchisqpow(128, wts, df, ncp, pow)
dvs <- dsumchisqpow(rvs, wts, df, ncp, pow)
qvs <- psumchisqpow(rvs, wts, df, ncp, pow)
pvs <- qsumchisqpow(ppoints(length(rvs)), wts, df, ncp, pow)
}
\seealso{
The upsilon distribution, 
\code{\link{dupsilon},\link{pupsilon},\link{qupsilon},\link{rupsilon}}.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{distribution}
