% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumgengamma.r
\name{sumgengamma}
\alias{dsumgengamma}
\alias{psumgengamma}
\alias{qsumgengamma}
\alias{rsumgengamma}
\alias{sumgengamma}
\title{The sum of generalized gammas distribution.}
\usage{
dsumgengamma(x, wts, a, d, pow, log = FALSE, order.max=6)

psumgengamma(q, wts, a, d, pow, lower.tail = TRUE, log.p = FALSE, order.max=6)

qsumgengamma(p, wts, a, d, pow, lower.tail = TRUE, log.p = FALSE, order.max=6)

rsumgengamma(n, wts, a, d, pow)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{wts}{the vector of weights.
This is recycled against the \code{a, d, pow}, but not against the \code{x,q,p,n}.}

\item{a}{the vector of the 'a' parameters of the variates.
This is recycled against the \code{wts, d, pow}, but not against the \code{x,q,p,n}.}

\item{d}{the vector of the 'd' parameters of the variates.
This is recycled against the \code{wts, a, pow}, but not against the \code{x,q,p,n}.}

\item{pow}{the vector of the 'p' parameters of the variates.
This is recycled against the \code{wts, a, d}, but not against the \code{x,q,p,n}.}

\item{log}{logical; if TRUE, densities \eqn{f} are given
as \eqn{\mbox{log}(f)}{log(f)}.}

\item{order.max}{the order to use in the approximate density,
distribution, and quantile computations, via the Gram-Charlier,
Edeworth, or Cornish-Fisher expansion.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log.p}{logical; if TRUE, probabilities p are given
as \eqn{\mbox{log}(p)}{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
 \eqn{P[X \le x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}
}
\value{
\code{dsumgengamma} gives the density, \code{psumgengamma} gives the
distribution function, \code{qsumgengamma} gives the quantile function,
and \code{rsumgengamma} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Density, distribution function, quantile function and random
generation for the distribution of the weighted sum of gamma
variates.
}
\details{
Let \eqn{X_i \sim G(a_i, d_i, p_i)}{X_i ~ G(a_i,d_i,p_i)}
be independently distributed generalized gamma variates with parameters
\eqn{a_i}{a_i}, \eqn{d_i}{d_i}, and inverse power \eqn{p_i}{p_i}.
Let \eqn{w_i} be given constants. Suppose
\deqn{Y = \sum_i w_i X_i.}{Y = sum w_i X_i.}
Then \eqn{Y}{Y} follows a weighted sum of generalized gammas distribution.
}
\note{
The PDF, CDF, and quantile function are approximated, via
the Edgeworth or Cornish Fisher approximations, which may
not be terribly accurate in the tails of the distribution.
You are warned.

The distribution parameters are \emph{not} recycled
with respect to the \code{x, p, q} or \code{n} parameters,
for, respectively, the density, distribution, quantile
and generation functions. This is for simplicity of
implementation and performance. It is, however, in contrast
to the usual R idiom for dpqr functions.
}
\examples{
wts <- c(1,-3,4)
a <- c(10,2,5)
d <- c(5,30,1)
pow <- c(1,2,1)
rvs <- rsumgengamma(128, wts, a, d, pow)
dvs <- dsumgengamma(rvs, wts, a, d, pow)
qvs <- psumgengamma(rvs, wts, a, d, pow)
pvs <- qsumgengamma(ppoints(length(rvs)), wts, a, d, pow)
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
Stacey, E. W. "A Generalization of the Gamma Distribution."
Annals of Mathematical Statistics 33, no. 3 (1962): 1187--1192.
\url{http://projecteuclid.org/euclid.aoms/1177704481}
}
\seealso{
The sum of log gen gammas distribution,
\code{\link{dsumloggengamma},\link{psumloggengamma},\link{qsumloggengamma},\link{rsumloggengamma}}.
}
\keyword{distribution}

