% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumchi.r
\name{sumchi}
\alias{dsumchi}
\alias{psumchi}
\alias{qsumchi}
\alias{rsumchi}
\alias{sumchi}
\title{The sum of (non-central) chis distribution.}
\usage{
dsumchi(x, wts, df, ncp=0, order.max=6, log = FALSE)

psumchi(q, wts, df, ncp=0, order.max=6, lower.tail = TRUE, log.p = FALSE)

qsumchi(p, wts, df, ncp=0, order.max=6, lower.tail = TRUE, log.p = FALSE)

rsumchi(n, wts, df, ncp=0)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{wts}{the vector of weights.
We do \emph{not} recycle this against \code{x,q,p,n}.}

\item{df}{the vector of degrees of freedom.
We do \emph{not} recycle this against \code{x,q,p,n}.}

\item{ncp}{the vector of non-centrality parameters.
We do \emph{not} recycle this against \code{x,q,p,n}.}

\item{order.max}{the order to use in the approximate density,
distribution, and quantile computations, via the Gram-Charlier,
Edeworth, or Cornish-Fisher expansion.}

\item{log}{logical; if TRUE, densities \eqn{f} are given
as \eqn{\mbox{log}(f)}{log(f)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log.p}{logical; if TRUE, probabilities p are given
as \eqn{\mbox{log}(p)}{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
 \eqn{P[X \le x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}
}
\value{
\code{dsumchi} gives the density, \code{psumchi} gives the
distribution function, \code{qsumchi} gives the quantile function,
and \code{rsumchi} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Density, distribution function, quantile function and random
generation for the distribution of the weighted sum of non-central
chis.
}
\details{
Let \eqn{X_i \sim \chi^2\left(\delta_i, \nu_i\right)}{X_i ~ chi^2(delta_i, v_i)}
be independently distributed non-central chi-squares, where \eqn{\nu_i}{v_i}
are the degrees of freedom, and \eqn{\delta_i}{delta_i} are the
non-centrality parameters.
Let \eqn{w_i} be given constants. Suppose
\deqn{Y = \sum_i w_i \sqrt{X_i}.}{Y = sum w_i sqrt(X_i).}
Then \eqn{Y}{Y} follows a weighted sum of chis distribution.
}
\note{
The PDF, CDF, and quantile function are approximated, via
the Edgeworth or Cornish Fisher approximations, which may
not be terribly accurate in the tails of the distribution.
You are warned.
}
\examples{
wts <- c(1,-3,4)
df <- c(100,20,10)
ncp <- c(0,2,1)
rvs <- rsumchi(128, wts, df, ncp)
dvs <- dsumchi(rvs, wts, df, ncp)
qvs <- psumchi(rvs, wts, df, ncp)
pvs <- qsumchi(ppoints(length(rvs)), wts, df, ncp)
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\seealso{
sum of chi-square distribution functions,
\code{\link{dsumchisq}, \link{psumchisq}, \link{qsumchisq}, \link{rsumchisq}},
}
\keyword{distribution}

