% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upsilon.r
\name{upsilon}
\alias{dupsilon}
\alias{pupsilon}
\alias{qupsilon}
\alias{rupsilon}
\alias{upsilon}
\title{The upsilon distribution.}
\usage{
dupsilon(x, df, t, log = FALSE, order.max=10)

pupsilon(q, df, t, lower.tail = TRUE, log.p = FALSE, order.max=10)

qupsilon(p, df, t, lower.tail = TRUE, log.p = FALSE, order.max=10)

rupsilon(n, df, t)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{df}{the degrees of freedom in the chi square. a vector. we do
\emph{not} vectorize over this variable.}

\item{t}{the scaling parameter on the chi. a vector. should be the same
length as \code{df}. we do \emph{not} vectorize over this variable.}

\item{log}{logical; if TRUE, densities \eqn{f} are given
as \eqn{\mbox{log}(f)}{log(f)}.}

\item{order.max}{the order to use in the approximate density,
distribution, and quantile computations, via the Gram-Charlier
series or the Cornish Fisher expansion.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log.p}{logical; if TRUE, probabilities p are given
as \eqn{\mbox{log}(p)}{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
 \eqn{P[X \le x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}
}
\value{
\code{dupsilon} gives the density, \code{pupsilon} gives the
distribution function, \code{qupsilon} gives the quantile function,
and \code{rupsilon} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Density, distribution function, quantile function and random
generation for the upsilon distribution.
}
\details{
Suppose \eqn{y_i \sim \chi^2\left(\nu_i\right)}{y_i ~ x^2(v_i)}
independently and independently of \eqn{Z}{Z}, a standard normal.
Then
\deqn{\Upsilon = Z + \sum_i t_i \sqrt{y_i/\nu_i}}{Y = Z + sum_i t_i sqrt(y_i/v_i)}
takes an upsilon distribution with parameter vectors
\eqn{[\nu_1, \nu_2, \ldots, \nu_k]', [t_1, t_2, ..., t_k]'}{<v_1, v_2, ..., v_k>, <t_1, t_2, ..., t_k>}.

The upsilon distribution is used in certain tests of
the Sharpe ratio for independent observations, and generalizes
the lambda prime random variable.
}
\note{
The PDF, CDF, and quantile function are approximated via
the Gram Charlier or Cornish Fisher approximations, which may
not be terribly accurate in the tails of the distribution.
}
\examples{
d1 <- dupsilon(1, 50, t=0.01)
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
Lecoutre, Bruno. "Another look at confidence intervals for the noncentral t distribution."
Journal of Modern Applied Statistical Methods 6, no. 1 (2007): 107--116.
\url{http://www.univ-rouen.fr/LMRS/Persopage/Lecoutre/telechargements/Lecoutre_Another_look-JMSAM2007_6(1).pdf}

Lecoutre, Bruno. "Two useful distributions for Bayesian predictive procedures under normal models."
Journal of Statistical Planning and Inference 79  (1999): 93--105.
}
\seealso{
lambda-prime distribution functions, \code{\link{dlambdap}, \link{plambdap}, \link{qlambdap}, \link{rlambdap}}.
}
\keyword{distribution}

