% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadists.r
\name{sadists-NEWS}
\alias{sadists-NEWS}
\title{News for package 'sadists'}
\description{
History of the 'sadists' package.

\newcommand{\sadists}{\href{https://github.com/shabbychef/sadists}}
}
\section{\sadists{} Version 0.2.6 (2026-07-16)}{

\itemize{
\item fix missing vignette references.
}
}

\section{\sadists{} Version 0.2.5 (2025-12-25)}{

\itemize{
\item fix issues with high df chisquare moments, closing #16
}
}

\section{\sadists{} Version 0.2.4 (2020-06-23)}{

\itemize{
\item CRAN emergency release.
}
}

\section{\sadists{} Version 0.2.3 (2017-03-19)}{

\itemize{
\item add product of normals distribution.
\item move github figures to location CRAN understands.
}
}

\section{\sadists{} Version 0.2.2 (2016-03-03) }{

\itemize{
\item work around bad \code{rchisq} when \eqn{df=0=ncp} (?)
\item incompatibilities in vignette with ggplot2 release.
}
}

\section{\sadists{} Version 0.2.1 (2015-06-12) }{

\itemize{
\item shiny app (h/t Dean Attali).
}
}

\section{\sadists{} Version 0.2.0 (2015-04-01) }{

\itemize{
\item add doubly non-central Beta and Eta distributions.
\item add (sum of) log chi-square distribution.
\item have products of chi-square depend on transform of this distribution.
}
}

\section{\sadists{} Initial Version 0.1.0 (2015-03-07) }{

\itemize{
\item first CRAN release.
}
}

