% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sadists.r
\name{sadists-NEWS}
\alias{sadists-NEWS}
\title{News for package 'sadists'}
\description{
History of the 'sadists' package.

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\sadists}{\CRANpkg{sadists}}
}
\section{\sadists{} Initial Version 0.1.0 (2015-03-07) }{

\itemize{
\item first CRAN release.
}
}

