% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sadists.r
\name{sadists-NEWS}
\alias{sadists-NEWS}
\title{News for package 'sadists'}
\description{
History of the 'sadists' package.

\newcommand{\sadists}{\href{https://github.com/shabbychef/sadists}}
}
\section{\sadists{} Version 0.2.1 (2015-06-12) }{

\itemize{
\item shiny app (h/t Dean Attali).
}
}

\section{\sadists{} Version 0.2.0 (2015-04-01) }{

\itemize{
\item add doubly non-central Beta and Eta distributions.
\item add (sum of) log chi-square distribution.
\item have products of chi-square depend on transform of this distribution.
}
}

\section{\sadists{} Initial Version 0.1.0 (2015-03-07) }{

\itemize{
\item first CRAN release.
}
}

